refactor(renderer): organize Markdown constructs
Move the event renderer and its tests out of the crate root into
construct-owned modules. Keep the public API and rendered output stable
while making lib.rs a concise facade.
